✅ test: give every package a tests/smoke and tests/unit folder
The smoke job introduced in the previous commit hardcoded a list of
specific test files in the workflow -- brittle, and it papered over two
packages (unxts.interop.gala, unxts.interop.matplotlib) that had no
public-API smoke check at all.

Every package (unxt itself and each packages/* member) now has:
  - tests/smoke/  -- fast, broad checks: imports, public-API re-exports
  - tests/unit/   -- everything else

Existing test_public_api.py files (already smoke tests in substance)
move wholesale where every test in them qualifies; unxts.api and
unxts.hypothesis get split, since each had one file mixing re-export
checks with real behavior/property-based tests. unxts.interop.gala and
unxts.interop.matplotlib get a new minimal smoke/test_public_api.py
(mirroring their siblings), and their existing behavior tests move to
unit/.

The CI smoke step now collects via `tests/smoke packages/*/tests/smoke`
instead of an enumerated file list -- a new package's smoke coverage is
just adding the folder, no workflow edit required. The full test-matrix
jobs and noxfile.py path resolution are untouched: they already pointed
at whole `tests/` directories, which transparently pick up the new
subfolders.
