+++ packages/unxts.hypothesis/tests/unit/test_strategies.py
@@ -1,19 +1,10 @@
-"""Smoke tests for the unxts.hypothesis public API."""
+"""Tests for the unxts.hypothesis strategies."""
 
 import pytest
 import unxts.hypothesis
 from hypothesis import given, strategies as st
 
 import unxt as u
-
-
-def test_all_symbols_present():
-    for name in unxts.hypothesis.__all__:
-        assert hasattr(unxts.hypothesis, name), f"unxts.hypothesis missing: {name}"
-
-
-def test_version_is_a_string():
-    assert isinstance(unxts.hypothesis.__version__, str)
 
 
 @given(q=unxts.hypothesis.quantities(unit="km/s"))
