+++ packages/unxts.linalg/tests/unit/test_edge_cases.py
