+++ packages/unxts.linalg/tests/unit/test_printing.py
