+++ packages/unxts.parametric/tests/unit/test_interchangeability.py
