+++ packages/unxts.parametric/tests/unit/test_pow_exponent.py
