+++ packages/unxts.parametric/tests/unit/test_printing.py
