+++ packages/unxts.parametric/tests/unit/test_quantity_staticvalue_eq.py
