+++ packages/unxts.parametric/tests/unit/test_static_value.py
