+++ pyproject.toml
@@ -446,7 +446,7 @@ ignore = [
 "**tests/**" = ["ANN", "D401", "E731", "INP001", "S101", "SLF001", "T20"]
 # `Qcls` is a class-valued fixture parametrized over the quantity classes; the
 # capitalized name (and using it as a constructor) is intentional.
-"**packages/unxts.parametric/tests/test_interchangeability.py" = [
+"**packages/unxts.parametric/tests/unit/test_interchangeability.py" = [
   "N802",
   "N803",
   "N806",
