+++ tests/smoke/__init__.py
@@ -1,0 +1,1 @@
+"""Smoke Tests."""
